backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.1.2
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
websocket-client==1.6.4
celery==5.3.4
redis==5.0.1
